app/requirements.txt
openai==1.74.1
langchain==0.3.25
langchain-community==0.3.23
langchain-openai==0.3.16
langchainhub==0.1.21
langsmith==0.3.42
langchain-core==0.3.58
tabulate==0.9.0
azure-identity==1.13.0
python-dotenv==1.1.0
azure-ai-documentintelligence==1.0.0b4
pypdf2==3.0.1
pypdf==3.15.0
python-docx==0.8.11
PyMuPDF==1.22.5
keras==3.5.0
numpy>=1.26.2,<2.2.0
pillow==10.0.0
tensorflow==2.19.0
streamlit==1.45.0
mammoth==1.4.2
docx2pdf==0.1.8